shared config using config server endpoints for prod env
