+++ packages/sygma-ui/src/configLoader/getSharedConfig.ts
@@ -1,5 +1,11 @@
-const getConfigFromConfigServer = () => {
-  return null;
+const CONFIG_SERVER_HOST = process.env.REACT_APP_CONFIG_SERVER_HOST;
+const CONFIG_SERVER_PORT = process.env.REACT_APP_CONFIG_SERVER_PORT;
+
+const getConfigFromConfigServer = async () => {
+  const response = await fetch(
+    `//${CONFIG_SERVER_HOST}:${CONFIG_SERVER_PORT}/share`
+  );
+  return await response.json();
 };
 
 const getLocalSharedConfig = async () => {
@@ -13,7 +19,7 @@ export async function getSharedConfig() {
   if (process.env.NODE_ENV === "production") {
     sharedConfig = await getConfigFromConfigServer();
   } else {
-    sharedConfig = getLocalSharedConfig();
+    sharedConfig = await getLocalSharedConfig();
   }
 
   return sharedConfig;
